Refactor AppRoutes and rename const.ts file
