+++ project/src/components/FilmCard/film-card.tsx
@@ -2,6 +2,7 @@ import {useState} from 'react';
 import {useNavigate} from 'react-router-dom';
 import {FilmType} from '../../types/film-type';
 import PromoPlayer from '../PromoPlayer/promo-player';
+import {AppRoutes} from '../../utils/const';
 
 type FilmCardProps = {
   film: FilmType
@@ -13,13 +14,13 @@ export default function FilmCard({film}: FilmCardProps): JSX.Element {
 
   return (
     <article className="small-film-card catalog__films-card" onMouseOver={() => setIsActive(true)}
-      onMouseOut={() => setIsActive(false)} onClick={() => navigate(`/films/${film.id}`)}
+      onMouseOut={() => setIsActive(false)} onClick={() => navigate(AppRoutes.FilmsRoot + film.id)}
     >
       <div className="small-film-card__image">
         <PromoPlayer film={film} cardActive={isActive}/>
       </div>
       <h3 className="small-film-card__title">
-        <a className="small-film-card__link" href="film-page.html">
+        <a className="small-film-card__link" href="#">
           {film.name}
         </a>
       </h3>
