+++ project/src/components/Footer/footer.tsx
@@ -1,10 +1,11 @@
 import {Link} from 'react-router-dom';
+import {AppRoutes} from '../../utils/const';
 
 export default function Footer() {
   return (
     <footer className="page-footer">
       <div className="logo">
-        <Link to={'/'} className="logo__link logo__link--light">
+        <Link to={AppRoutes.Root} className="logo__link logo__link--light">
           <span className="logo__letter logo__letter--1">W</span>
           <span className="logo__letter logo__letter--2">T</span>
           <span className="logo__letter logo__letter--3">W</span>
