+++ project/src/components/Header/header.tsx
@@ -1,6 +1,6 @@
 import {Link, useNavigate} from 'react-router-dom';
 import {useAppSelector} from '../../hooks/store-hooks';
-import {AuthorizationStatus} from '../../utils/consts';
+import {AppRoutes, AuthorizationStatus} from '../../utils/const';
 import {dispatch} from '../../types/state';
 import { logoutAction } from '../../store/aip-actions';
 
@@ -13,7 +13,7 @@ export default function Header() {
   return (
     <header className="page-header film-card__head">
       <div className="logo">
-        <Link className="logo__link" to={'/'}>
+        <Link className="logo__link" to={AppRoutes.Root}>
           <span className="logo__letter logo__letter--1">W</span>
           <span className="logo__letter logo__letter--2">T</span>
           <span className="logo__letter logo__letter--3">W</span>
@@ -25,7 +25,7 @@ export default function Header() {
           ?
           <ul className="user-block">
             <li className="user-block__item">
-              <div className="user-block__avatar" onClick={() => navigate('/myList')}>
+              <div className="user-block__avatar" onClick={() => navigate(AppRoutes.MyList)}>
                 <img
                   src="img/avatar.jpg"
                   alt="User avatar"
@@ -35,12 +35,12 @@ export default function Header() {
               </div>
             </li>
             <li className="user-block__item">
-              <Link className="user-block__link" to={'/login'} onClick={() => dispatch(logoutAction())}>Sign out</Link>
+              <Link className="user-block__link" to={AppRoutes.Login} onClick={() => dispatch(logoutAction())}>Sign out</Link>
             </li>
           </ul>
           :
           <div className="user-block">
-            <Link className="user-block__link" to={'/login'}>Sign in</Link>
+            <Link className="user-block__link" to={AppRoutes.Login}>Sign in</Link>
           </div>
       }
