+++ project/src/components/app/app.tsx
@@ -1,5 +1,5 @@
 import {BrowserRouter, Routes, Route} from 'react-router-dom';
-import {AppRoutes, AppStatus, AuthorizationStatus} from '../../utils/consts';
+import {AppRoutes, AppStatus, AuthorizationStatus} from '../../utils/const';
 
 import MainPage from '../../pages/MainPage/main-page';
 import LoginPage from '../../pages/LoginPage/login-page';
